Converst HTML tables into markdown
Most of the tables were converted except the one in permissions-role-resources.md
This table uses links and is kind of index. This index can be converted into some kind of list one day.
As long as it is a table - it is easier to maintain it as a HTML table (it uses <br> and confitions) than
one line in markdown table