Clarify cancel order process (#944)
* Added Cancel an order section

Although it is explained in the first paragraph of this topic, added a new section to make it more prominent.

* Update orders.md

Added links to related topics to better connect the dots.

* Reworked Orders topic

This content duplicated the workflow topic. I rewrote the topic to give an overview of the Orders grid and as an into to the process.