Update Usage data Collection
This pull request (PR) is to update the Admin Usage data collection section. With the integration of gainsight (in-product guidance) within Magento Admin starting with 2.4.2 and onwards, and the fact that users only experience this value add if they consent to opt-in for Admin Usage collection, we want to update the opt-in admin usage section in the User Guides to include value information about in-product guidance.